Lock single bits in bitmap
Locking a bit instead of a byte avoids potential deadlock on free when
two transactions need the same two bytes in the bitmap but they lock
them in different order.

We can optimize allocation by handing out a range of freebits if all
of them are free, instead of 1 bit at the time.