3 versions of smallfile
one runs inside server's process, one runs as an NFS client, and one
run as a file system client.  this makes it easier to analyze where
performance is going.

modify all so that they delete the file on each iteration. smallfile.c
didn't do that, probably because delayed commit would absorbed all of
it.