Small tweaks and fix a few bugs
Support partial writes correctly: don't return immediately out of
space, but the number of bytes written.  The next write will fail with
out of space.

Don't release a locked bit during a scan of the bitmap if the
transaction already had the bit locked. This happens when there is
only space in the file system for a partial write.

Put inodes on CommitFh too

Allow larger file names so that git clone works

Set atime/mtime on inode creation

Cloning xv6 repo and building it on goose-nfsd works