Checkpoint a simpler and better shrinking plan
Inode has a new field ShrinkSize, which Setattr sets to the old size
of an inode, when truncating the inode to a new smaller size.  With
this information any thread can shrink the inode, and in incremental
steps. Like before, the truncate starts a shrinker thread, which
shrinks the inode in several transactions, if shrinking modifies more
blocks than fit in the wal.

If a later Write/SetAttr regrows a file while shrinking is in
progress, it first helps with shrinking the file before growing
it. This avoid that later reads see old data.  The old plan didn't
have a solution for this later case.

This plans handles crashes nicely too.  On recovery, nothing must be
done: when a later write/setattr wants to grow the inode and sees that
it hasn't been shrunk yet, it will shrink it.