Checkpoint simplifying allocator
Don't keep locks for individual bits in bitmap (removing addr-to-lock
map) and don't read bits from disk (removing disk reads from
allocator). Instead, read bitmap blocks once at startup. Allocator
maintains invariant that in-memory allocator state (set of free object
IDs) is a subset of the on-disk allocator state (set of free bits in
bitmap). Allocation returns one of the in-memory IDs, and marks it
in-use after allocating an object ID.  Commits adds a write to the
allocated bitmap bit.  Same for free.

Todo: fix abort (return allocated bits to in-memory state), but this
commit exposed a latent bug in mkdir: a re-used inode might be still
be in progress of being shrunk, and mkdir doesn't help/wait.