Help shrinking in doCreate when creating directory or link
Unsatisfactory solution: there is a risk of inode loss.  helpShrinker
may start a new transaction and a crash may happen before the name is
added to the parent directory in the next transaction.  AllocInode,
Writing to its first block (for directory and links), and adding the
name to the directory should be atomic.  Not clear how to achieve this
when the allocated inode must be shrunk first.