Checkpoint shrinking an inode when allocating it
When allocating a free node that is in the process of being shrunk,
mkdir/symlink want to write it (e.g., for . and ..).  Since shrinking
and writing might not fit in one transaction, doCreate helps shrinking
the inode and then restarts allocating an inode.  (We could also have
bumped the size and zero the blocks that it needs, as in earlier
versions, but that seems more complicated.)