Remove Done() from cache
The cache can evict an item at any point, for example, even when an
inode is in use by a transaction.  That is ok, because a transaction
has the inode locked and reads in that transaction observe the last
write to the inode in the transaction.  Once the inode is unlocked,
other transactions can lock the inode and then read it from the cache,
which will read it from disk. That read will include the changes from
the last transaction because they are written into the log before the
transaction releases the inode.

This plan makes the cache purely a performance optimization.